Added player stop function and camera stop following on collider trigger
The player character stops for 5 seconds and the camera does not follow. Optionally, the camera follows player again after timer too.

To do this, I made the JustWalk function public and added a boolean parameter, then called parts of it in DoggoCollider.cs to control player speed. Did similar with CameraControl.cs to call main camera, so it can be forced to no longer follow player.